refactor: 디렉터리 변경 및 애니메이션 제거
