+++ frontend/src/Router.tsx
@@ -3,13 +3,13 @@ import { ROUTER_PATH } from './constants';
 
 import CustomerList from './pages/Admin/CustomerList';
 import ManageCafe from './pages/Admin/ManageCafe';
-import CouponList from './pages/CouponList';
+import CouponList from './pages/Customer/CouponList';
 import EnterPhoneNumber from './pages/Admin/EnterPhoneNumber';
-import Login from './pages/Login';
+import Login from './pages/Customer/Login';
 import SignUp from './pages/SignUp';
 import NotFound from './pages/NotFound';
 import RegisterCafe from './pages/Admin/RegisterCafe';
-import MyPage from './pages/MyPage';
+import MyPage from './pages/Customer/MyPage';
 import Template from './components/Template';
 import CustomCouponDesign from './pages/Admin/CouponDesign/CustomCouponDesign';
 import ModifyCouponPolicy from './pages/Admin/ModifyCouponPolicy';
@@ -17,15 +17,16 @@ import SelectCoupon from './pages/Admin/EarnStamp/SelectCoupon';
 import RewardPage from './pages/Admin/RewardPage';
 import EarnStamp from './pages/Admin/EarnStamp';
 import CustomerTemplate from './components/Template/CustomerTemplate';
-import RewardList from './pages/RewardList';
-import RewardHistoryPage from './pages/HistoryPage/RewardHistory';
-import StampHistoryPage from './pages/HistoryPage/StampHistory';
+import RewardList from './pages/Customer/RewardList';
+import RewardHistoryPage from './pages/Customer/HistoryPage/RewardHistory';
+import StampHistoryPage from './pages/Customer/HistoryPage/StampHistory';
 import Auth from './pages/Auth';
 import AdminLogin from './pages/Admin/AdminLogin';
 import AdminAuth from './pages/Admin/AdminAuth';
 import TemplateCouponDesign from './pages/Admin/CouponDesign/TemplateCouponDesign';
-import InputPhoneNumber from './pages/InputPhoneNumber';
+import InputPhoneNumber from './pages/Admin/InputPhoneNumber';
 import CustomerNotFound from './pages/NotFound/CustomerNotFound';
+import Main from './pages/Customer/Main';
 
 const AdminRoot = () => {
   return (
@@ -82,7 +83,7 @@ const Router = () => {
       element: <CustomerRoot />,
       errorElement: <CustomerNotFound />,
       children: [
-        { index: true, element: <CouponList /> },
+        { index: true, element: <Main /> },
         { path: ROUTER_PATH.auth, element: <Auth /> },
         { path: ROUTER_PATH.login, element: <Login /> },
         { path: ROUTER_PATH.signup, element: <SignUp /> },
