+++ frontend/src/pages/Customer/CouponList/Coupon/index.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import { Coupon as CouponType } from '../../../types';
+import { Coupon as CouponType } from '../../../../types';
 import { CouponWrapper, ImageForLoading } from './style';
-import CouponLoading from '../../../assets/coupon_load_img_for_customer.png';
+import CouponLoading from '../../../../assets/coupon_load_img_for_customer.png';
 
 interface CouponProps {
   coupon: CouponType;
