+++ frontend/src/pages/Customer/CouponList/CouponDetail/index.tsx
@@ -1,5 +1,5 @@
 import FlippedCoupon from '../FlippedCoupon';
-import Text from '../../../components/Text';
+import Text from '../../../../components/Text';
 import {
   CafeImage,
   CloseButton,
@@ -11,8 +11,8 @@ import {
 import { BiArrowBack } from 'react-icons/bi';
 import { FaRegClock, FaPhoneAlt, FaRegBell, FaRegTrashAlt } from 'react-icons/fa';
 import { FaLocationDot } from 'react-icons/fa6';
-import { Coupon } from '../../../types';
-import { parsePhoneNumber } from '../../../utils';
+import { Coupon } from '../../../../types';
+import { parsePhoneNumber } from '../../../../utils';
 import {
   QueryObserverResult,
   RefetchOptions,
@@ -20,12 +20,12 @@ import {
   useMutation,
   useQuery,
 } from '@tanstack/react-query';
-import { CouponRes } from '../../../types/api';
-import { getCafeInfo } from '../../../api/get';
-import { deleteCoupon } from '../../../api/delete';
-import useModal from '../../../hooks/useModal';
-import Alert from '../../../components/Alert';
-import CustomerLoadingSpinner from '../../../components/LoadingSpinner/CustomerLoadingSpinner';
+import { CouponRes } from '../../../../types/api';
+import { getCafeInfo } from '../../../../api/get';
+import { deleteCoupon } from '../../../../api/delete';
+import useModal from '../../../../hooks/useModal';
+import Alert from '../../../../components/Alert';
+import CustomerLoadingSpinner from '../../../../components/LoadingSpinner/CustomerLoadingSpinner';
 
 interface CouponDetailProps {
   isDetail: boolean;
