+++ frontend/src/pages/Customer/HistoryPage/RewardHistory/index.tsx
@@ -1,11 +1,11 @@
 import { useQuery } from '@tanstack/react-query';
-import { getMyRewards } from '../../../api/get';
+import { getMyRewards } from '../../../../api/get';
 import { CafeName, HistoryItem, DateTitle, HistoryList } from '../style';
-import { parseStringDateToKorean, sortMapByKey, transformEntries } from '../../../utils';
-import { RewardHistoryDateProperties, RewardHistoryType } from '../../../types';
-import { DATE_PARSE_OPTION } from '../../../constants';
+import { parseStringDateToKorean, sortMapByKey, transformEntries } from '../../../../utils';
+import { RewardHistoryDateProperties, RewardHistoryType } from '../../../../types';
+import { DATE_PARSE_OPTION } from '../../../../constants';
 import HistoryPage from '../HistoryPage';
-import CustomerLoadingSpinner from '../../../components/LoadingSpinner/CustomerLoadingSpinner';
+import CustomerLoadingSpinner from '../../../../components/LoadingSpinner/CustomerLoadingSpinner';
 
 export const concatHistoryDate = (
   reward: RewardHistoryType,
