+++ frontend/src/pages/Customer/Main/index.tsx
@@ -1,0 +1,7 @@
+import CouponList from '../CouponList';
+
+const Main = () => {
+  return <CouponList />;
+};
+
+export default Main;
