+++ frontend/src/pages/Customer/RewardList/style.tsx
