+++ frontend/src/style/keyframes.ts
@@ -11,17 +11,6 @@ export const swap = keyframes`
   }
 `;
 
-export const detail = keyframes`
-  0% {
-    transform: translateY(15px) scale(1.05);
-    display: flex;
-  }
-  
-  100%{
-    transform: translateY(-219%) scale(0.86);
-  }
-`;
-
 export const popup = keyframes`
   0% {
     transform: translateY(100%);
